refactor: organize role / tenant functions to respective files
